Added rule to allow activation of the standard fn keys
This allows activation of standard function keys with fn button
on per profile basis. For now it is not possible, because the
activation of F1, F2... buttons is a global settings. So settings
the F1, F2 and so on buttons to F1, F2... and using this rule the
inversion of the behaviour is possible on per profile basis.