Add helpful rules for the Matias Ergo Pro Keyboard + personal rules
The [Matias Ergo Pro](https://matias.ca/ergopro/) is one of the few
ergonomic mechanical keyboards on the market with an mostly-standard
Mac layout. Two quirks stand out:

1. The navigation keys (`home`, `end`, `page_up`, and `page_down`) are
where the `right_option` key is on a MacBook or Magic Keyboard.
2. The `right_control` key is where the B key *would* be *if* it were
on the right side of the split (which it isn't).

The following rules remap theys keys in a context-sensitive manner:

1. Change navigation keys to right_option if pressed with another key
2. Change right_control to b if pressed alone, or while typing

This commit also introduces some personal rules of mine, as well as the
`KeyCodes` and `Variable` utility classes used to build both rule sets
(and likely to be helpful for other future rules too).