Disabled lazy option
If the lazy option is true, clicking a link while pressing
the Command key on Google Chrome doesn't open the link in a new tab.

Additionally, the lazy option was needed to enable key mapping change
when the key in "from" and another key are pressed at the same time only.
But it is not needed now because "to_if_alone" works instead of that.