Add NuPhy NuType-F1 keyboard changes
More information about this device:
https://nuphy.com

This change allows the keyboard to work more like expected experience.
In kickstarter page:
https://www.kickstarter.com/projects/nuphy/nutype-revolutionizing-the-laptop-typing-experience/description
this was the original intent, but then modified before shipping the
units to production:
> For example, you can press 'FN' + '~' to substitute the use of 'ESC' button.

When writting markdown it is normal to use '`' often to add some code
blocks. Having to press 'FN' to get the normal key to work doesn't seem
natural and doesn't work like other keyboards that have the 'ESC' button
in a different location (as expected - like new MacBook Pro 16").