Add Archiss Quattro TKL JIS layout
More information about this device
https://www.archisite.co.jp/products/archiss/quattro/tkl-jp/

Quattro TKL JIS is a Japanese keyboard device for PC.
This keyboard has a unique key layout that the left and right of the space key replace to mouse buttons.
But the Japanese general key layout for macOS has international keys  "Kana" and "Eisu" at the left and right of the space key. These necessary keys have been stolen, so that keyboard is very difficult to use as Japanese macOS users.

This rule recovers and replace the international keys for Quattro TKL JIS.