Add files via upload
This one I am particularly proud of. Mainly because I desperately missed an app, I used to use years ago with iTunes, which kept iTunes and system volume in sync. While I just prefer this behaviour generally, it also enables you to control non-system-wide AirPlay audio via the volume keys, which I need quite often.
Still missing is key repeat functionality, as this is apparently not supported for shell commands. Therefor it is intentionally disabled altogether (for the volume keys) to not behave weirdly.