Fix Neo2 mod 4 key on Apple keyboards
non_us_backslash and grave_accent_and_tilde are swapped on Apple and
Windows keyboards and need to be handled differently. This does not just
affect the left mod 4 key, but also the key left of '1', especially the
dead key functions with layer 4 and 6.

To get the handling right, I merged the modifier key rules and layer
4 rule. There is probably not really a use case to only have one of
  those rules activated anyways.

Fixes #1057.