Improve "backtick_z_slash_as_ctrl" modification ease typing "Z"
This improvement posts "Z" immediately if shift and z are pressed
together. This makes it easier to type "Z" quickly, instead of getting
"z" because by the time the z keypress is posted on keyup, the shift key
has already been released. With this modification, it's still possible
to type ctrl-shift-SOMEKEY, you just need to press z first and then
shift, as if you press shift first and then z, you will get keyrepeat of
z. I think ctrl-shift is a rare enough key combo that this is a good
tradeoff.

On my keyboard layout (Dvorak) ;/: is on the QWERTY z key, so this was
tripping me up a lot while trying to use Vim (typing colon enters the
command mode).