Create f18.json
Add a short and simple example of remapping a single function key or other key to a key with modifiers.  This example is the default for Capture with SnagIt so mapping it to an extra function key on an external keyboard.  This is a better example to start with for many purposes than the complicated ones for other Hyper keys.